backend/requirements.txt
fastapi
uvicorn
supabase
httpx
beautifulsoup4
python-dotenv
pydantic
bcrypt
